Fix some styling issue should follow max width to container 1220px Fix pagination active issue
