+++ components/Header/index.tsx
@@ -2,6 +2,7 @@ import { HamburgerIcon } from '@chakra-ui/icons';
 import { Box, Button, Container, Flex, IconButton, Image, Spacer, Stack, useDisclosure } from '@chakra-ui/react';
 import Cookies from 'js-cookie';
 import Link from 'next/link';
+import { useRouter } from 'next/router';
 import { FC } from 'react';
 import Logo from '../../assets/icons/logo';
 
@@ -9,10 +10,11 @@ const Header: FC = (props) => {
   const { isOpen, onOpen, onClose } = useDisclosure();
   const handleToggle = (): void => (isOpen ? onClose() : onOpen());
   const token = Cookies.get('token');
+  const router = useRouter();
 
   return (
     <Flex
-      as="nav"
+      as="header"
       align="center"
       flexShrink={0}
       flexDirection="row"
@@ -26,7 +28,7 @@ const Header: FC = (props) => {
       bg="#FFF"
       {...props}
     >
-      <Container position="static" display="flex" maxW="container.xl">
+      <Container height="100%" position="static" display="flex" maxW="1220px" padding={0}>
         <Flex align="center" mr={5}>
           <Link href={`/`}>
             <IconButton height="auto" variant="unstyled" aria-label="logo-icon">
@@ -47,7 +49,12 @@ const Header: FC = (props) => {
           flexGrow={1}
           mt={{ base: 4, md: 0 }}
         >
-          <Button variant="unstyled">
+          <Button
+            height="100%"
+            variant="unstyled"
+            borderBottom={router.pathname == '/' || router.pathname.includes('product') ? '2px solid #6366F1' : ''}
+            borderRadius="0"
+          >
             <Link href={`/`}>Products</Link>
           </Button>
           <Spacer />
