+++ pages/products/add-product.tsx
@@ -1,21 +1,24 @@
 import { Box } from '@chakra-ui/react';
 import Breadcrumbs from '@components/Breadcrumbs';
-import Layout from '@components/Layout';
+import ContentLayout from '@components/Layout/ContentLayout';
+import MainLayout from '@components/Layout/MainLayout';
 import AddEditProduct from '@modules/AddEditProduct.tsx';
 import Head from 'next/head';
 import { FC } from 'react';
 
 const AddProduct: FC = (): JSX.Element => {
   return (
-    <Layout>
-      <Head>
-        <title>Add Product</title>
-      </Head>
-      <Box w="100%" padding="90px 110px">
-        <Breadcrumbs />
-        <AddEditProduct />
-      </Box>
-    </Layout>
+    <MainLayout>
+      <ContentLayout>
+        <Head>
+          <title>Add Product</title>
+        </Head>
+        <Box w="100%">
+          <Breadcrumbs />
+          <AddEditProduct />
+        </Box>
+      </ContentLayout>
+    </MainLayout>
   );
 };
